pyproject.toml
[project]
name = "studybot"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Golubchikov",email = "anton1programmist@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "python-telegram-bot[job-queue] (>=22.4,<23.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "studybot", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.13.0"
mypy = "^1.17.1"
types-aiofiles = "^24.1.0.20250822"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 79

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

